Backend/ml/requirements.txt
# Python Recommendation System Requirements
# Compatible with Python 3.8 - 3.12+
# Install with: pip install -r requirements.txt

# Core ML Libraries (Python 3.12+ compatible)
numpy>=1.26.0,<2.0.0
scipy>=1.11.0,<2.0.0
scikit-learn>=1.3.0,<2.0.0

# Deep Learning (Optional - for GNN)
# Uncomment if you want to use GNN recommender
torch>=2.1.0
torch-geometric>=2.4.0

# MongoDB
pymongo>=4.6.0,<5.0.0

# Environment Management
python-dotenv>=1.0.0,<2.0.0

# Data Processing
pandas>=2.1.0,<3.0.0

# Optional: For better performance
joblib>=1.3.0,<2.0.0  # For model serialization

# Note: Using flexible version ranges for better compatibility
# The system works without torch (GNN will be disabled)
redis==4.6.0   # For caching (if you want Python-side caching)
